requirements.txt
allure-pytest==2.14.2
allure-python-commons==2.14.2
attrs==25.3.0
backports.functools-lru-cache==2.0.0
certifi==2025.4.26
charset-normalizer==3.4.2
future==1.0.0
h11==0.16.0
idna==3.10
iniconfig==2.1.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
outcome==1.3.0.post0
packaging==25.0
pluggy==1.5.0
PySocks==1.7.1
pytest==8.3.5
python-dotenv==1.1.0
referencing==0.36.2
requests==2.32.3
rpds-py==0.25.1
selene==2.0.0rc9
selenium==4.32.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.13.2
urllib3==2.4.0
utils==1.0.2
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0